RGB raw color sensor simulator support (#1023)
* set-rgbraw-mode-circle-color

* preliminary_sim_rgb_raw_support

* updateRGBWheel.ts

* preliminary-changes

Implemented toggle switches for each RGB component.

* preliminary-changes-2

* final-changes

* rounding-change

So that there is no difference in uart values, as well as values in the label of the toggle switches.

* sim-color-mode-set-defl-val

Prior to this, the color mode was not taken into account. It turns out that the initial value is 50. There is no such index in the array. showPorts didn't output anything for this sensor initially until you chose some color yourself in the simulator. If you look at what the value is there, then 50 is displayed. Therefore, I made the default value 0 - nothing.