implementation of treason via getUartStatus
I figured out that the getUartStatus() method on input gives status 8 when the sensor has changed mode most of the time. Displayed in control.dmesg and checked it, but periodically the value jumped more than 1024 as before, the value was 40 thousand. I additionally added a condition that the value cannot be greater than 1023.