Changes in mode change for color sensor (#1025)
* Update color.ts

When changing the mode to another, the new values are checked to make sure they are in the range that the mode corresponds to. We also additionally skip a few values to make sure that the sensor gives us values from the set mode.

* algorithm-improvements

Maximum cycle time for protection. Do not work for a port where no sensor is connected.

* Update color.ts

* Update color.ts

* Update color.ts

* implementation of treason via getUartStatus

I figured out that the getUartStatus() method on input gives status 8 when the sensor has changed mode most of the time. Displayed in control.dmesg and checked it, but periodically the value jumped more than 1024 as before, the value was 40 thousand. I additionally added a condition that the value cannot be greater than 1023.

* Update color.ts

* Update color.ts

* Update color.ts

* Update libs/color-sensor/color.ts

Co-authored-by: Joey Wunderlich <jwunderl@users.noreply.github.com>

* Update libs/color-sensor/color.ts

Co-authored-by: Joey Wunderlich <jwunderl@users.noreply.github.com>

* return ColorCal

---------

Co-authored-by: Joey Wunderlich <jwunderl@users.noreply.github.com>