Editor display changes (#1036)
* Update style.less

The first change returns the background pattern as in the stable version.

The second change is to fix #downloadArea overlapping on the simulator. I don’t really understand why this bug is visible only in pxt-ev3. This bug is not detected in microbit, but the #downloadArea field is also larger than the height of #editortools itself. I compared the properties, but they seem to be all the same.

* revet-editortools-overflow-hidden

* Update style.less

Filling the background color is not required, because The #editortools parent element already has a background fill.