Pxt core bump 9.1.1x and more (#1038)
* bump-pxt-core

Timely update of the pxt-core version to support the project.

* active-item-fill-height-change

* pxt-core-bump-9.1.14

* pxt-bump-9-1-15

* add-subcategories-planned-sens

* ht-color-sen-add-lib

* Remove-queryArr-and-infoArr

Got rid of queryArr and infoArr, which I created to support the color sensor in the RawRgb mode in the simulator.

* revert-ms-automation-ext

Revert automation extension.

* pxt-bump-9.1.16

* move-status-indicator-to-new-group

Move the status indicator to a separate group. It was in a button group, but according to the ev3 manual it is not a button, so I created a separate group for it and output the code to a separate file. It's logical.

* exitProgram-moved-from-buttons-to-program

exitProgram has been moved from buttons to the program group.

* bump-pxt-9.1.17

* pxt-bump-9.1.18

* pxt-bump-9.1.19

* i2c-sens-work-when-start-program

The i2c sensors will work when the program starts. Before this, the i2c sensor started working only when other sensors were reconnected, calling detectDevices() again.

* Update input.ts

* Update input.ts

* pxt-bump-9.1.21

* pxt-bump-9.1.23

* color-for-separator-lego-ana-microsoft-logos

* Update style.less

* Update style.less

* revert-approvedRepos

* Update targetconfig.json