Nxt Light Sensor support (#1027)
* add-subcategories

Add NXT Sensors subcategorie

* first-support-version

The light sensor gives real values. Mode change not supported...

* update

* Update ns.ts

* sens-sim-support

* min-and-max-metod

* sim-support-new-chenges

* for-sim

The simulator now understands that it has an EV3 or NXT analog sensor enabled. The devType value of the virtual sensor will be passed so that the sensor becomes active in the simulator and transmits its values to the simulator.

* Update input.ts

* add-test

* fix-convert-range-bug

* Update light.ts

Setting the range of values for determining black and white in reflection mode and in ambient light mode is now different.

* Update test.ts

* Update test.ts

* Update test.ts

* removing-function-doesnt-work

I'm removing an unnecessary function that doesn't work, which I took from the pelikhan change. It does not apply to this change with the sensor and data input.

* checking-input-values

* set-nxt-light-sensor-svg

* Update light.ts

* update-for-sim

* Update NXT Light Sensor.svg

* resolving-conflict-with-master

* resolving-conflict-with-master-2

* resolving-conflict-with-master-3

* Update nxtLightSensorView.ts

* Update input.ts

* value-range-update

These numbers were obtained by testing on 4 sensors. The black value is the sensor aimed at the void, and the light value is aimed at the white Lego brick.

* Update light.ts

* some-changes-for-the-mode

* LightWheelControl-is-always-activated-except-in-none-mode

LightWheelControl is always activated except in NONE mode. Otherwise, the mode was activated immediately when the sensor was turned on in the simulator, when the operating mode had not yet been activated.

* 4096-to-4095

Fix range - 0..4095

* enum-NXTLightIntensityMode

* light-sensor-svg-updage

* update-light-sensor-ts

* ambient-modes-set-invisible

Disable the visibility of ambient blocks so that they are not used, because There is no implementation of disabling LED lighting.

* range-from-lego-sources

https://github.com/mindboards/ev3sources-xtended/blob/master/ev3sources/lms2012/lms2012/Linux_AM1808/sys/settings/typedata.rcf

* sim-fix-for-sensor

Solving the problem that the field in the simulator and the value from the sensor on the screen could be different.
And one more thing...

* query-update-and-add-treshold-blocks

* test-upd

* Update lightWheel.ts

* changes-light-sensor-for-sim

The changes are aimed at supporting reflection and lighting modes, not raw modes.

* deviceType-analog-sens-class

* Update light.ts

* fix-range-for-refLight-and-ambLight

* export-enum-DevConOff

* bug-fix-from-previous-version

* sensor-activation

Setting the mode so that the sensor starts working in the simulator. Otherwise, it, like a touch sensor, does not create view control, because in the touch sensor this was not necessary. Without this change, lightView would only activate if a sensor with type uart was used in the code.

* Update light.ts

Support blocks have been removed, which may be available later. The ambient mode has been removed, because There is no implementation of turning off the backlight LED. Now it is always on.

* Update light.ts

* lib-not-include-by-defl

Make sure that the library is not included by default. It will need to be enabled via extensions.

* light-to-bright

* add-docs

* Update nxt-light-sensor.md