Update input.ts
When we find the DAL.CONN_NXT_DUMB sensor, we additionally call setAnalogMode(), which passes the sensor mode to the simulator. Perhaps the same thing needs to be done for DAL.CONN_INPUT_DUMB, but there is only one analog sensor ev3 (touch sensor), and it works without this change.